Split pkg/sharing/sharing.go in several files
